Changed limit to accept long instead of int and added short-circuit to avoid creating nested empty streams.

+++ library/src/main/java/au/id/ajlane/iostreams/IOStream.java
@@ -229,7 +229,7 @@ public interface IOStream<T> extends Closeable
      *
      * @return A limited view of the stream.
      */
-    default IOStream<T> limit(final int size)
+    default IOStream<T> limit(final long size)
     {
         return IOStreams.limit(this, size);
     }
